move JavaScript to own dir, add unicodeMaps
