+++ packages/quicktype-core/src/language/JavaScript/unicodeMaps.ts
